Use now registered OutputCollectors package
