+++ Project.toml
@@ -19,6 +19,7 @@ Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 LoggingExtras = "e6f89c97-d47a-5376-807f-9c37f3926c36"
 MbedTLS = "739be429-bea8-5141-9913-cc70e7f3736d"
 ObjectFile = "d8793406-e978-5875-9003-1fc021f44a92"
+OutputCollectors = "6c11c7d4-943b-4e2b-80de-f2cfc2930a8c"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 PkgLicenses = "fc669557-7ec9-5e45-bca9-462afbc28879"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
